fixed bug reported by saurabh when multiple questions in for loop
Earlier - when multiple questions were in a for loop, at the next
iteration the question from the previous iteration were showing up
again. This was because of incorrect allcation of questionNoIndex_

What we have done is re-assigned the questionNoIndex_ at the time of
pushing into question_list. I think this should solve the problem of
this type permananently. Need to test a case with 3 questions inside a
for loop to be sure.

More test cases tomorrow