stage 1 of language translation
our_ui.js works upto named attributes.

However, the question title has to be re-factored in all cases of the
display function.

It has to be extracted out in the case of single question on screen and
multiple question on screen.

Also it should be made independent of rq/nq/video_q questions