exported all image assets in the script to a file
the file is named as json but it is not currently in json format.

Dont Think we need the file in json format, all that we will need at a
later stage is to verify that all the assets listed in this file have
indeed been uploaded by the survey scripter. Otherwise the survey will
go live without the image/media and will cause a bad user experience.

This is basically a facility to help him check that the survey is in
order - like a lint