compiling to java ... beginnings
at last they day has come when I have to bite the Java bullet. I knew I
would have to get into it and today we make a start.

The emscripten compilation to javascript is too slow for my liking.
Everything works, but end consumers would not like an app that ran so
slow. Hence, we try to compile to native java which can be used on
android.

The step to android is in 2 phases
1. get the code to compile in java
2. add an android skin around the java compiled code